+++ src/microservices/customer/graphql/input/index.ts
@@ -1,0 +1,2 @@
+export * from './create-customer-profile.input';
+export * from './update-customer-profile.input';
