+++ src/microservices/customer/graphql/models/index.ts
@@ -1,0 +1,1 @@
+export * from './customer-gql.model';
