+++ src/microservices/customer/graphql/mutations/customer-mutation.resolver.ts
@@ -2,10 +2,9 @@ import { Args, Mutation, Resolver } from '@nestjs/graphql';
 import { UseGuards } from '@nestjs/common';
 
 import { GqlJwtAuthGuard } from '@auth/graphql/guards';
-import { GqlCustomer } from '../models/customer-gql.model';
+import { GqlCustomer } from '../models';
 import { CustomerService } from '../../services';
-import { CreateCustomerProfileInput } from '../input/create-customer-profile.input';
-import { UpdateCustomerProfileInput } from '../input/update-customer-profile.input';
+import { CreateCustomerProfileInput, UpdateCustomerProfileInput } from '../input';
 
 @Resolver((of) => GqlCustomer)
 export class CustomerMutationResolver {
